% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_worms_get_aphia_id.R
\name{api_worms_get_aphia_id}
\alias{api_worms_get_aphia_id}
\title{Query WORMS API for Aphia ID from species identification name}
\usage{
api_worms_get_aphia_id(scientific_names)
}
\arguments{
\item{scientific_names}{}
}
\value{
a dataframe with 3 columns: taxonomic ID (Aphia ID), scientific_name, and name_to_evaluate (the string used to query WORMS API)
}
\description{
Wrapper around functions from worrms package with special functionality
to deal with issues commonly seen with MarineGEO IDs (sp., spp., A, B, C, etc.)
and to format the output
}
\examples{
phylum <- "Mollusca" # 51
api_worms_get_aphia_id(phylum)
}
