% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_worms_get_classification.R
\name{api_worms_get_classification}
\alias{api_worms_get_classification}
\title{Query WORMS API for taxonomic levels with Aphia ID}
\usage{
api_worms_get_classification(taxonomic_ids)
}
\arguments{
\item{taxonomic_ids}{}
}
\value{
A dataframe with a column for \code{scientific_name} and \code{taxonomic_id} with accompanying columns for each taxon rank.
}
\description{
Returns all associated taxon ranks for a Aphia ID (e.g., kingdom, phylum, genus, species).
}
\examples{
phylum <- 51 # Mollusca
api_worms_get_classification(phylum)
}
