% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_taxonomy.R
\name{db_get_taxonomy_by_taxonomic_id}
\alias{db_get_taxonomy_by_taxonomic_id}
\title{Get taxonomic classifications by  taxonomic ID}
\usage{
db_get_taxonomy_by_taxonomic_id(taxonomic_ids, taxonomic_levels = NULL)
}
\arguments{
\item{taxonomic_levels}{Vector of classification levels. By default, returns all available levels}

\item{scientific_names}{Vector of taxonomic IDs to get classifications for}
}
\value{
dataframe of taxonomic classifications and taxonomic ID (usually Aphia ID)
}
\description{
Get taxonomic classifications by  taxonomic ID
}
