% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_marinegeo_L2.R
\name{db_marinegeo_L2}
\alias{db_marinegeo_L2}
\title{Connect to a controlled data directory with Arrow. Target should be defined in MarineGEO data structure. Use "dataframe" as return type to load entire dataset into R. Default, "arrow", allows you to filter dataset before collecting into R.}
\usage{
db_marinegeo_L2(table_id, return_type = "arrow")
}
\arguments{
\item{return_type}{"arrow" to return an active binding to directory or "dataframe" to collect all data at connection}
}
\description{
Connect to a controlled data directory with Arrow. Target should be defined in MarineGEO data structure. Use "dataframe" as return type to load entire dataset into R. Default, "arrow", allows you to filter dataset before collecting into R.
}
\examples{
table_id <- "reef-life-survey-data-marinegeo-v1"
ds <- db_marinegeo_L2(table_id, "arrow")
ds \%>\% filter(site_code == "BC4") \%>\% collect()
}
