% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_check_point_counts.R
\name{qc_check_point_counts}
\alias{qc_check_point_counts}
\title{Check that the total number of point counts for each quadrat totals the number in "points in quadrat"}
\usage{
qc_check_point_counts(data)
}
\arguments{
\item{data}{A data frame to validate.}
}
\value{
A \link{qc_issues} tibble with one row per quadrat-transect combination that violates the point-count rules.
}
\description{
Validates that the total number of primary point counts for a given
quadrat adds up to the number of points in the quadrat
Designed for use in MarineGEO QA/QC pipelines and can be called
directly or via \code{\link[=qc_run]{qc_run()}}.
}
\examples{
data <- data.frame(
 transect = c(1,1,1,1,1,2,2,2,2,2),
  cover_type = c("live_oyster", "box_oyster", "cultch","hash","sediment","live_oyster", "box_oyster", "cultch","hash","sediment" ),
  quadrat = c(5,5,5,5,5,5,5,5,5,5),
  point_count = c(9,13,20,34,4,9,13,20,34,5),
  points_in_quadrat= c(81,81,81,81,81,81,81,81,81,81)
)

output_2 <- qc_check_point_counts(df_out)
}
