% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_controlled_vocabulary.R
\name{qc_mg_categorical_values}
\alias{qc_mg_categorical_values}
\title{Evaluate categorical values
Evaluate based on lookup table stored with R package}
\usage{
qc_mg_categorical_values(df, table_id)
}
\arguments{
\item{table_id}{}
}
\value{
vector of row numbers for values that are not in a controlled vocabulary
}
\description{
Evaluate categorical values
Evaluate based on lookup table stored with R package
}
