% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_controlled_vocabulary.R
\name{qc_mg_categorical_values_details}
\alias{qc_mg_categorical_values_details}
\title{Evaluate categorical values for column details
Evaluate based on lookup table stored with R package}
\usage{
qc_mg_categorical_values_details(df, table_id)
}
\arguments{
\item{table_id}{}
}
\value{
table with details about which columns are being flagged on each row for the invalid categorical values.
}
\description{
Evaluate categorical values for column details
Evaluate based on lookup table stored with R package
}
