% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_column_data_types.R
\name{qc_mg_column_data_types}
\alias{qc_mg_column_data_types}
\title{Verify that data types in table match arrow data types}
\usage{
qc_mg_column_data_types(df, table_id)
}
\arguments{
\item{table_id}{}
}
\value{
df
}
\description{
If there is a mismatch, an error is created with information on the offending column name and data type.
The function is designed to fit into the select all columns and write_csv pipe,
so it returns the original dataframe if it passes the tests. The error will prevent mismatch
types from getting written to a directory used for Apache Arrow.
}
