% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_missing_taxonomic_id.R
\name{qc_mg_missing_taxonomic_id}
\alias{qc_mg_missing_taxonomic_id}
\title{Check for observations missing a taxonomic ID}
\usage{
qc_mg_missing_taxonomic_id(df)
}
\arguments{
\item{df}{}
}
\value{
a list of row numbers that have an NA or non-numeric value in the "taxonomic_id" column
}
\description{
Check that no observations are missing a value in the "taxonomic_id" column.
Typically these values are Aphia IDs. Negative numbers are MarineGEO-specific IDs.
}
\examples{
qc_mg_missing_taxonomic_id(df)
}
