% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_controlled_vocabulary.R
\name{qc_mg_partner_codes}
\alias{qc_mg_partner_codes}
\title{Check for registered partner codes}
\usage{
qc_mg_partner_codes(df)
}
\arguments{
\item{df}{}
}
\value{
vector of row numbers for partner codes that are not in a controlled vocabulary
}
\description{
Check for registered partner codes
}
