% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_mg_controlled_vocabulary.R
\name{qc_mg_site_names}
\alias{qc_mg_site_names}
\title{Check for registered site names}
\usage{
qc_mg_site_names(df)
}
\arguments{
\item{df}{}
}
\value{
vector of row numbers for site names that are not in a controlled vocabulary
}
\description{
Check for registered site names
}
