% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rls_empty_metadata.R
\name{qc_rls_empty_metadata}
\alias{qc_rls_empty_metadata}
\title{QC Survey Summary Table}
\usage{
qc_rls_empty_metadata(df)
}
\arguments{
\item{df}{}
}
\value{
a list of rows in the summary data table to highlight because they have a metadata issue that need to be corrected.
}
\description{
Looks for NA and Blank values in the dataframe summary table
}
\examples{
filepath <- "inst/extdata/test_rls_data_EPA_one_header_row.xlsx"
df <- utl_rls_load_excel(filepath)
summary_df <- utl_rls_summarize_table(df)
dives_missing_metadata <- qc_rls_empty_metadata(summary_df)
}
