% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rls_fish_count_over_100.R
\name{qc_rls_fish_count_over_100}
\alias{qc_rls_fish_count_over_100}
\title{Check for fish count over 100}
\usage{
qc_rls_fish_count_over_100(df)
}
\arguments{
\item{df}{}
}
\value{
a list of row numbers that are fish species with a count over 100
}
\description{
Flag any rows where phylum is "Chordata" and "Total" is > 100.
These rows should be reviewed, but a flag does not mean it's an error.
}
\examples{
qc_rls_fish_count_over_100(df)
}
