% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rls_inverts_with_sizes.R
\name{qc_rls_inverts_with_sizes}
\alias{qc_rls_inverts_with_sizes}
\title{Check for invertebrates that have been assigned size counts}
\usage{
qc_rls_inverts_with_sizes(df)
}
\arguments{
\item{df}{}
}
\value{
a list of row numbers that are invert species with a size assigned to them.
}
\description{
Check that the rls dataset is a dataframe and contains columns \code{Total}, \code{Inverts}, \code{phylum}, and all the size count columns
finds rows where the phylum is not "Chordata" and any of the size coulmns have a numerical value over 0.
}
\examples{
rows_inverts_with_sizes <- qc_rls_inverts_with_sizes(df)
}
