% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rls_total_count_0.R
\name{qc_rls_total_count_0}
\alias{qc_rls_total_count_0}
\title{Check for total count of 0}
\usage{
qc_rls_total_count_0(df)
}
\arguments{
\item{df}{}
}
\value{
a list of row numbers where Total equals 0
}
\description{
Flag any row where the Total count is 0
}
\examples{
qc_rls_total_count_0(df)
}
