% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rls_verts_with_invert_count.R
\name{qc_rls_verts_with_invert_count}
\alias{qc_rls_verts_with_invert_count}
\title{Check for vertebrates that have been given an invert count}
\usage{
qc_rls_verts_with_invert_count(df)
}
\arguments{
\item{df}{}
}
\value{
a list of row numbers that are vertebrae species with an invert count assigned to them.
}
\description{
Check that the rls dataset is a dataframe and contains columns \code{Inverts} and \code{phylum}
finds rows where the phylum is "Chordata" and the invert count column has a numerical value over 0.
}
\examples{
rows_verts_with_invert_count <- qc_rls_verts_with_invert_count(df)
}
