% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_rls_deprecated.R
\name{utl_join_taxonomy_by_scientific_name_deprecated}
\alias{utl_join_taxonomy_by_scientific_name_deprecated}
\title{Join taxonomic classification to input dataframe}
\usage{
utl_join_taxonomy_by_scientific_name_deprecated(
  df,
  identification_column_name = "scientific_name",
  taxonomic_levels = NULL
)
}
\arguments{
\item{df}{A data frame to join taxonomic levels to}

\item{identification_column_name}{Defaults to "scientific_name"}

\item{taxonomic_levels}{Vector of classification levels. By default, returns all options}
}
\value{
A dataframe with taxonomic levels as columns
}
\description{
The function calls \code{utl_get_taxonomy_by_scientific_name_deprecated()} to query the MarineGEO taxonomic
database by scientific name. If the sample is identified by a column not named "scientific_name",
it should be defined in the \code{identification_column_name} argument.
}
\examples{
\dontrun{
df <- utl_mg_load_excel_deprecated(
  "inst/extdata/test_rls_data_EPA_one_header_row.xlsx",
  "reef-life-survey-data-marinegeo-v1",
  "DATA"
) |>
  dplyr::filter(Method == "1" | Method == "2")

utl_join_taxonomy_by_scientific_name_deprecated(
  df,
  identification_column_name = "Species"
)

utl_join_taxonomy_by_scientific_name_deprecated(
  df,
  identification_column_name = "Species",
  taxonomic_levels = "phylum"
)
}
}
