% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_data_resources.R
\name{utl_mg_data_index}
\alias{utl_mg_data_index}
\title{Return the MarineGEO data index table}
\usage{
utl_mg_data_index()
}
\value{
df: a dataframe containing the MarineGEO data index
}
\description{
Return the MarineGEO data index table
}
\examples{
utl_mg_data_index()
}
