% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_data_resources.R
\name{utl_mg_data_structure}
\alias{utl_mg_data_structure}
\title{Return the MarineGEO data table column and data type schema}
\usage{
utl_mg_data_structure()
}
\value{
df: a dataframe containing the MarineGEO data structures
}
\description{
Return the MarineGEO data table column and data type schema
}
\examples{
utl_mg_data_structure()
}
