% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_load_excel.R
\name{utl_mg_load_excel}
\alias{utl_mg_load_excel}
\title{Load MarineGEO Excel data into R}
\usage{
utl_mg_load_excel(filepath, output_table, sheet_name)
}
\arguments{
\item{sheet_name}{}
}
\value{
a dataframe
}
\description{
Load one or more sheets from a MarineGEO Excel file into R.The \code{output_table}
argument is used to select the correct sheet from the Excel file.
}
