% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_data_resources.R
\name{utl_mg_table_relationships}
\alias{utl_mg_table_relationships}
\title{Return the MarineGEO table relationships}
\usage{
utl_mg_table_relationships()
}
\value{
df: a dataframe containing the MarineGEO table relationships
}
\description{
Return the MarineGEO table relationships
}
\examples{
utl_mg_table_relationships()
}
