% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_rls_fill_missing_metadata.R
\name{utl_rls_fill_missing_metadata}
\alias{utl_rls_fill_missing_metadata}
\title{Filling missing metadata columns based on other processed RLS data}
\usage{
utl_rls_fill_missing_metadata(df)
}
\arguments{
\item{df}{}
}
\description{
For columns "vis", "direction", "time", and "photoquadrats", the function
converts "BLANK" to NA and checks processed L2 data for fill values from
the same survey.
}
