% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_rls_deprecated.R
\name{utl_rls_fill_missing_metadata_deprecated}
\alias{utl_rls_fill_missing_metadata_deprecated}
\title{Filling missing metadata columns based on other processed RLS data}
\usage{
utl_rls_fill_missing_metadata_deprecated(df)
}
\arguments{
\item{df}{the data frame to check for missing metadata.}
}
\value{
dataframe with all metadata
}
\description{
For columns "vis", "direction", "time", and "photoquadrats", the function
converts "BLANK" to NA and checks processed L2 data for fill values from
the same survey.
}
