% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_rls_load_excel.R
\name{utl_rls_load_excel}
\alias{utl_rls_load_excel}
\title{Load Reef Life Survey Excel into R}
\usage{
utl_rls_load_excel(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
A dataframe without any header rows or with empty trailing rows.
}
\description{
Load the DATA sheet from a RLS (Reef Life Survey) Excel file into R.
Drops first and potentially second row based on contents of \code{Method} column.
Drops trailing empty rows based on \code{Total} column
}
\examples{
filepath <- "inst/extdata/test_rls_data_EPA_two_header_rows.xlsx"
filepath <- "inst/extdata/test_rls_data_EPA_one_header_row.xlsx"
df <- utl_rls_load_excel(filepath)
}
