% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_rls_summarize_table.R
\name{utl_rls_sample_event_summary}
\alias{utl_rls_sample_event_summary}
\title{Create a survey summary table}
\usage{
utl_rls_sample_event_summary(df)
}
\arguments{
\item{df}{}
}
\value{
a list of unique date, depth, vis, and sitecode combinations with other dive-metadata variables and True-False statements for method-Block combinations
}
\description{
lists the unique survey metadata combinations. Check for the existance of method/block 1 and 2.
}
\examples{
filepath <- "inst/extdata/test_rls_data_EPA_two_header_rows.xlsx"
df <- utl_rls_load_excel(filepath)
summary_df <- utl_rls_summarize_table(df)
}
