% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_rls_deprecated.R
\name{db_get_taxonomy_by_scientific_name_deprecated}
\alias{db_get_taxonomy_by_scientific_name_deprecated}
\title{Get taxonomic classifications by scientific name}
\usage{
db_get_taxonomy_by_scientific_name_deprecated(
  scientific_names,
  taxonomic_levels = NULL
)
}
\arguments{
\item{scientific_names}{Vector of scientific names to get taxonomic classifications for}

\item{taxonomic_levels}{Vector of classification levels. By default, returns all available levels}
}
\value{
dataframe of taxonomic classifications and taxonomic ID (usually Aphia ID)
}
\description{
Get taxonomic classifications by scientific name
}
\examples{
utl_get_taxonomy_by_scientific_name(c("Gastropoda", "Dardanus venosus"))
utl_get_taxonomy_by_scientific_name("Dardanus venosus", taxonomic_levels = "phylum")
utl_get_taxonomy_by_scientific_name("Dardanus venosus", taxonomic_levels = "pilum") # misspelling
}
