% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_table_wrangler.R
\name{shiny_launch_table_wrangler}
\alias{shiny_launch_table_wrangler}
\title{Launch the MarineGEO Table Wrangler Shiny application}
\usage{
shiny_launch_table_wrangler()
}
\description{
Launch the MarineGEO Table Wrangler Shiny application
}
