% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utl_join_taxonomy_by_scientific_name_deprecated.R
\name{utl_join_taxonomy_by_scientific_name_deprecated}
\alias{utl_join_taxonomy_by_scientific_name_deprecated}
\title{Join taxonomic classification to input dataframe}
\usage{
utl_join_taxonomy_by_scientific_name_deprecated(
  df,
  identification_column_name = "scientific_name",
  taxonomic_levels = NULL
)
}
\arguments{
\item{df}{A data frame to join taxonomic levels to}

\item{identification_column_name}{Defaults to "scientific_name"}

\item{taxonomic_levels}{Vector of classification levels. By default, returns all options}
}
\value{
A dataframe with taxonomic levels as columns
}
\description{
The function calls \code{utl_get_taxonomy_by_scientific_name()} to query the MarineGEO taxonomic
database by scientific name. If the sample is identified by a column not named "scientific_name",
it should be defined in the \code{identification_column_name} argument.
}
\examples{
df <- utl_rls_load_excel("inst/extdata/test_rls_data_EPA.xlsx", sheet = "DATA") |>
 dplyr::filter(Method == "1" | Method == "2")
utl_join_taxonomy_by_scientific_name(df, identification_column_name = "Species")
utl_join_taxonomy_by_scientific_name(df, identification_column_name = "Species", "phylum")
}
