% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_data_resources.R
\name{utl_mg_column_order}
\alias{utl_mg_column_order}
\title{Return the column order for a MarineGEO table}
\usage{
utl_mg_column_order(table_id)
}
\arguments{
\item{table_id}{}
}
\value{
a character vector representing column names
}
\description{
Return the column order for a MarineGEO table
}
