% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_rls_deprecated.R
\name{utl_mg_load_excel_deprecated}
\alias{utl_mg_load_excel_deprecated}
\title{Load MarineGEO Excel data into R}
\usage{
utl_mg_load_excel_deprecated(filepath, output_table, sheet_name)
}
\arguments{
\item{filepath}{the path to the rls file being loaded}

\item{output_table}{the name of the intended L2 final product}

\item{sheet_name}{the name of the sheet int he excel file where the data can be found.}
}
\value{
a dataframe
}
\description{
Load one or more sheets from a MarineGEO Excel file into R.The \code{output_table}
argument is used to select the correct sheet from the Excel file.
}
\examples{
\dontrun{
utl_mg_load_excel_deprecated(
  "L1/rls_example.xlsx",
  "reef-life-survey-data-marinegeo-v1",
  "DATA"
)
}
}
