% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_mg_backfill.R
\name{utl_sav_backfill_cover}
\alias{utl_sav_backfill_cover}
\title{Backfill seagrass cover data
Ensures absence (cover = 0) row present for each species observed at a site}
\usage{
utl_sav_backfill_cover(df)
}
\arguments{
\item{df}{}
}
\description{
Backfill seagrass cover data
Ensures absence (cover = 0) row present for each species observed at a site
}
